Refactor tests to separate files
