+++ test/constants.js
@@ -1,0 +1,13 @@
+/**
+ * Created by JiaHao on 10/11/15.
+ */
+
+const TEST_URLS = ['http://apple.com', 'http://jiahaog.github.io/ajax-test-page/'];
+
+module.exports = {
+    TIMEOUT: 30000,
+    TEST_URLS: TEST_URLS,
+    AJAX_TEST_PAGE: TEST_URLS[1],
+    AJAX_TEST_PAGE_BUTTON: 'http://jiahaog.github.io/ajax-test-page/ajax_button_test.html',
+    INVALID_URL: 'http://insdasjdlkas.com/'
+};
